+++ panels/pose_tools.py
@@ -21,4 +21,5 @@ class RexTools3PoseToolsPanel(bpy.types.Panel):
         layout.separator()
         
         col = layout.column()
+        col.operator("rextools3.init_weight_paint", text="Init Weight Paint", icon='WPAINT_HLT')
         col.operator("rextools3.setup_pose_copier", text="Setup Pose Copier", icon='POSE_HLT')